gcs-video-automations-v2/requirements_web.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.8.0
google-cloud-storage==2.10.0
google-generativeai==0.3.2
Pillow>=11.0.0
google-cloud-texttospeech==2.16.5
